hypothesis: get lawful ancestors using the `laws()` classmethod.
Did not know about this earlier.

Move in the tests to `test_laws_resolution.py`, since they are testing
`laws`, not anything about `hypothesis`.
